Meaningless code cleanup
I made an effort to go through and remove a lot of the reasons for the
copious warnings that gcc throws out when compiling this project. Most
of those are related to comparison of signed vs. unsigned integers; and
most of those were in looping constructs (for int i=0;..etc). The main
reason I made the effort is in order for the more serious warnings that
actually can cause problems to be more visible. There are a few of those
around, mostly originating from a somewhat silly idea of mine to use
object memory addresses as identifiers, which might have been ok, except
that these will be a different size depdending on whether a computer is
32 or 64 bit (or whatever the future might bring). These don't actually
matter at the moment, since I never actually made any use of them. Still
at least the warnings for those more serious issues are not lost in a
mass of other warnings. To be completely frank, I'm still somewhat confused
as to when it is good to use unsigned and signed values, and this is
something that still bites me from time to time.

Naturally this caused some problems since I didn't perhaps do it carefully
enough. Hopefully I've caught all the problems caused by these changes,
but time will tell.

+++ annotator/annotationViewer.cpp
@@ -75,7 +75,7 @@ void AnnotationViewer::updateWidgets(probe_data* pdata){
   //  clearWidgets();
   //cout << "cleared Widgets, and am about to make some new ones" << endl;
   int myId = client->userId;
-  for(int i=0; i < pdata->annotations.size(); i++){
+  for(uint i=0; i < pdata->annotations.size(); i++){
     /// ok we'll need to find out what our identity is later, but for now assume we can edit.. 
     /// and we'll need to connect some signals and stuff here later on..
     // use the client pointer to determine what user id I have.. and so forth..
@@ -98,7 +98,7 @@ void AnnotationViewer::updateSessionWidgets(probe_data* pdata){
   //clearWidgets();
   //cout << "cleared Widgets, and am about to make some new ones" << endl;
   int myId = client->userId;
-  for(int i=0; i < pdata->sessions.size(); i++){
+  for(uint i=0; i < pdata->sessions.size(); i++){
     /// ok we'll need to find out what our identity is later, but for now assume we can edit.. 
     /// and we'll need to connect some signals and stuff here later on..
     // use the client pointer to determine what user id I have.. and so forth..
@@ -140,12 +140,12 @@ void AnnotationViewer::addLabel(QString labelString){
   
     
 void AnnotationViewer::clearWidgets(){
-  for(int i=0; i < widgets.size(); i++){
+  for(uint i=0; i < widgets.size(); i++){
     widgets[i]->hide();
     delete widgets[i];
   }
   widgets.resize(0);
-  for(int i=0; i < labels.size(); i++){
+  for(uint i=0; i < labels.size(); i++){
     labels[i]->hide();
     delete labels[i];
   }
@@ -163,11 +163,11 @@ void AnnotationViewer::clearWidgets(){
 void AnnotationViewer::resizeEvent(QResizeEvent* e){
   // need to resize all the widgets. 
   frame->resize(scroll->viewport()->width(), frame->height());
-  for(int i=0; i < widgets.size(); i++){
+  for(uint i=0; i < widgets.size(); i++){
     widgets[i]->setFixedWidth(scroll->viewport()->width());
     widgets[i]->setFixedHeight(widgets[i]->sizeHint().height());
   }
-  for(int i=0; i < labels.size(); i++){
+  for(uint i=0; i < labels.size(); i++){
     labels[i]->setFixedWidth(scroll->viewport()->width());
   }
   //update();
@@ -177,7 +177,7 @@ void AnnotationViewer::includeGenes(bool include){
   AnnotationWidget* aw = (AnnotationWidget*)sender();
   map<int, int>::iterator it;
   if(include){
-    for(int i=0; i < aw->annotation.otherMembers.size(); i++){
+    for(uint i=0; i < aw->annotation.otherMembers.size(); i++){
       it = includedGenes.find(aw->annotation.otherMembers[i]);
       if(it == includedGenes.end()){
 	includedGenes.insert(make_pair(aw->annotation.otherMembers[i], 1));
@@ -186,7 +186,7 @@ void AnnotationViewer::includeGenes(bool include){
       }
     }
   }else{
-    for(int i=0; i < aw->annotation.otherMembers.size(); i++){
+    for(uint i=0; i < aw->annotation.otherMembers.size(); i++){
       it = includedGenes.find(aw->annotation.otherMembers[i]);
       if((*it).second > 1){
 	(*it).second--;
@@ -199,7 +199,7 @@ void AnnotationViewer::includeGenes(bool include){
   // more than one, -it doesn't make much sense to load the current one
   loadIncluded->setDisabled(includedGenes.size() < 2);
   QString setSize;
-  setSize.sprintf("Set size : %d", includedGenes.size());
+  setSize.sprintf("Set size : %d", (int)includedGenes.size());
   includedGeneCount->setText(setSize);
   //cout << "Number of Genes Included : " << includedGenes.size() << endl;
 }
